corrections mineurs

+++ gestionEtudiant/src/main/java/com/tp2/gestionEtudiant/repository/FeuilleDePresenceRepository.java
@@ -11,6 +11,6 @@ public interface FeuilleDePresenceRepository  extends CrudRepository<FeuilleDePr
     
     public List<FeuilleDePresence> findByMailEtudiant (String mailEtudiant);
 
-    public FeuilleDePresence findById(int id);
+    public FeuilleDePresence findById(long id);
 
 }
